+++ CHANGELOG.md
@@ -4,3 +4,4 @@ Changelog
 ## [Unreleased]
 ### Changed
 - [ECS Cluster] Update to Amazon 2 ECS optimized AMI.
+- [ECS Cluster] Use custom AMI based on Amazon 2 ECS optimized that includes SSM.